% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/carscore.batch.R
\name{carscore.batch}
\alias{carscore.batch}
\title{Computation of CAR scores over a matrix of predictands}
\usage{
carscore.batch(predictand.matrix, predictor.matrix)
}
\arguments{
\item{predictand.matrix}{A matrix of predictand (response variables). Columns are predictands}

\item{predictor.matrix}{A matrix of predictors. Columns are predictors, number of rows must equal number of layers in \code{predictand.matrix}.}
}
\value{
A matrix of carscores.
}
\description{
Takes a matrix of predictands and a matrix of predictors (each predictand/predictor gets a column)
and returns a list of CAR scores, one for each columns in the predictand.matrix.
}

