% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carscore_batch.R
\name{carscore_batch}
\alias{carscore_batch}
\title{Computation of CAR scores over a matrix of predictands}
\usage{
carscore_batch(predictand.matrix, predictor.matrix)
}
\arguments{
\item{predictand.matrix}{A matrix of predictand (response variables). Columns are predictands}

\item{predictor.matrix}{A matrix of predictors. Columns are predictors, number of rows must equal number of layers in \code{predictand.matrix}.}
}
\value{
A list of carscores, one for each columns in the predictand.matrix.
}
\description{
Takes a matrix of predictands and a matrix of predictors (each predictand/predictor gets a column)
and returns a list of CAR scores, one for each columns in the predictand.matrix.
}

