% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_mlm.R
\name{CV_mlm}
\alias{CV_mlm}
\title{Computation of LOOCV and AICc for mlm objects}
\usage{
CV_mlm(obj)
}
\arguments{
\item{obj}{An object of class mlm or lm.}
}
\value{
A matrix of of CV and AICc values for each model.
}
\description{
Extends the CV method from the forecast package to handle 
multi-predictand linear models (objects of class mlm). Specifically, it calculates
the leave-one-out cross validataion statistic (also known as the PRESS statistic) and
the corrected Akaike's Information Criterion.
}

