% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\title{Remove the duplicate entries in a matrix-like object}
\usage{
duplicates(mat)
}
\arguments{
\item{mat}{A matrix or matrix-like object such as a data.frame or data.table.}
}
\value{
A named list containing a matrix of the unique rows of the matrix and
a vector matching the original indices to rows in the unique matrix.
}
\description{
This is a convenient function that removes the duplicated entries (rows)
in a matrix, but keeps a map of which of the original indices
match the duplicates.
}

