% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/errors.paleoCAR.models.batch.R
\name{errors.paleocar.models.batch}
\alias{errors.paleocar.models.batch}
\title{Get a RasterBrick of reconstruction LOOCV errors}
\usage{
errors.paleocar.models.batch(models, prediction.years = NULL)
}
\arguments{
\item{models}{A PaleoCAR batch model, as returned from \link{paleoCAR.models.batch}.}

\item{prediction.years}{The set of years over which to generate error rasters.}
}
\value{
A RasterBrick containing the LOOCV statistics per year.
}
\description{
This extracts the leave-one-out cross-validation (LOOCV) error information
from a set of PaleoCAR models and represents them as a RasterBrick with
layers for each reconstructed year. The LOOCV statistic is the sum of cross-validated
squared errors for the model over the calibration period.
}

