% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_paleocar_models_batch.R
\name{errors_paleocar_models_batch}
\alias{errors_paleocar_models_batch}
\title{Get a RasterBrick of reconstruction LOOCV errors}
\usage{
errors_paleocar_models_batch(models, prediction.years = NULL)
}
\arguments{
\item{models}{A PaleoCAR batch model, as returned from \link{paleocar_models_batch}.}

\item{prediction.years}{The set of years over which to generate error rasters.}
}
\value{
A RasterBrick containing the LOOCV statistics per year.
}
\description{
This extracts the leave-one-out cross-validation (LOOCV) error information 
from a set of PaleoCAR models and represents them as a RasterBrick with
layers for each reconstructed year. The LOOCV statistic is the sum of cross-validated 
squared errors for the model over the calibration period. This method returns the average errors
by dividing the LOOCV statistic by the number of predictand observations, and then taking the square root.
Resulting error is delivered in the units of the predictand.
}

