% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictor_matrix.R
\name{get_predictor_matrix}
\alias{get_predictor_matrix}
\title{Get a matrix of predictors for a calibration}
\usage{
get_predictor_matrix(chronologies, calibration.years, min.width = NULL)
}
\arguments{
\item{chronologies}{A matrix of predictors to be subset. Should be output from FedData::get_itrdb().
Columns are predictors, rows are years.}

\item{calibration.years}{A vector of years to be used for calibration}

\item{min.width}{integer, indicating the minimum number of tree-ring samples allowed for that year of a chronology to be valid.}
}
\value{
A matrix of predictors for a calibration
}
\description{
This subsets a chronologies matrix to a set of calibration years.
}

