% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPredictorMatrix.R
\name{getPredictorMatrix}
\alias{getPredictorMatrix}
\title{Get a matrix of predictors for a calibration}
\usage{
getPredictorMatrix(chronologies, calibration.years)
}
\arguments{
\item{chronologies}{A matrix of predictors to be subset.
Columns are predictors, rows are years.}

\item{reconstruction.years}{A vector of years to be used for calibration}
}
\value{
A matrix of predictors for a calibration
}
\description{
This subsets a chronologies matrix to a set of calibration years.
}

