% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{itrdb}
\alias{itrdb}
\title{Data from the International Tree Ring Data Bank}
\format{list
\describe{
  \item{metadata}{A data.table or SpatialPointsDataFrame (if makeSpatial==TRUE) of the 
  locations and names of extracted ITRDB chrononlogies}
  \item{widths}{A matrix of tree-ring widths/densities given user selection}
  \item{depths}{A matrix of tree-ring sample depths}
}}
\source{
\url{https://www.ncdc.noaa.gov/data-access/paleoclimatology-data/datasets/tree-ring}
}
\usage{
itrdb
}
\description{
A dataset containing the tree-ring data from the ITRDB for 10-degree buffer around 
Mesa Verde National Park.
}
\details{
These data were extracted from the ITRDB using the \code{FedData::get_itrdb()} function.
}
\keyword{datasets}

