% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeMonotonic.R
\name{makeMonotonic}
\alias{makeMonotonic}
\title{Make a vector strictly monotonic}
\usage{
makeMonotonic(vect, decreasing = T)
}
\arguments{
\item{vect}{A numeric vector.}

\item{decreasing}{Should the output indicate monotonically decreasing?}
}
\value{
A logical vector.
}
\description{
Given a vector \code{v} of length greater than 1, returns a logical vector
the same length as \code{v} indicating whether each element of \code{v} is
greater (or less-than, if \code{decreasing==TRUE}) than all previous elements.
}

