% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrix.expand.R
\name{matrix.expand}
\alias{matrix.expand}
\title{Expand a matrix, filling-down rows as needed}
\usage{
matrix.expand(the.matrix, new.rows)
}
\arguments{
\item{the.matrix}{A numeric matrix.}

\item{new.rows}{A character vector of row names,
of which a subset whould be the row names of \code{the.matrix}.}
}
\value{
A numeric matrix.
}
\description{
Given a matrix \code{m} and a vector \code{v} of new row IDs, returns a new matrix
with \code{length(v)} rows, and fills the empty rows with data from the nearest preceding row
containing data.
}

