% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvnp}
\alias{mvnp}
\title{MVNP Spatial Polygon}
\format{
SpatialPolygonsDataFrame
}
\source{
\url{http://nrdata.nps.gov/programs/Lands/meve_tracts.zip}
}
\usage{
mvnp
}
\description{
A dataset containing the spatial polygon defining the boundary 
of Mesa Verde National Park in southwestern Colorado.
}
\keyword{datasets}
