% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvnp_prism}
\alias{mvnp_prism}
\title{PRISM water_year precipitation}
\format{
RasterStack
}
\source{
\url{http://www.prism.oregonstate.edu/}
}
\usage{
mvnp_prism
}
\description{
A dataset containing the water-year (October--September) gridded precipitation estimates for 
Mesa Verde National Park derived from the 800m PRISM dataset.
}
\keyword{datasets}
