% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleocar_batch.R
\name{paleocar_batch}
\alias{paleocar_batch}
\title{Fit PaleoCAR models to a RasterBrick of predictands, and generate reconstruction, error, and model size RasterBricks}
\usage{
paleocar_batch(chronologies, predictands, calibration.years,
  prediction.years = NULL, label, out.dir = "./OUTPUT/", min.width = NULL,
  meanVar = "none", floor = NULL, ceiling = NULL, asInt = F,
  force.redo = F, verbose = F, generate.reconstruction = T,
  return.objects = T)
}
\arguments{
\item{chronologies}{An ITRDB object, as in from FedData::get_itrdb.}

\item{predictands}{A RasterBrick or RasterStack of the numeric predictand (response) variable.}

\item{calibration.years}{An integer vector of years corresponding to the layers in the \code{predictands} brick.}

\item{prediction.years}{An optional integer vector of years for the reconstruction.
If missing, defaults to the total years present in \code{chronologies}.}

\item{label}{A character label for the reconstruction, for saving.}

\item{out.dir}{The directory to which output is to be saved.}

\item{min.width}{integer, indicating the minimum number of tree-ring samples allowed for that year of a chronology to be valid.}

\item{meanVar}{A character string indicating the type of mean-variance matching to perform: either "none" (default), "calibration", or "chained".}

\item{floor}{Numeric, an optional lower bound for reconstructed values, such as \code{0} for precipitation reconstructions.}

\item{ceiling}{Numeric, an optional upper bound for reconstructed values.}

\item{asInt}{Logical, should reconstructed values be rounded to integers for saving?}

\item{force.redo}{Logical, should all computations be re-computed?}

\item{verbose}{Logical, display status messages during run.}

\item{generate.reconstruction}{Logical, should the reconstruction be generated (as opposed to only the models)?}

\item{return.objects}{Logical, should objects be returned (as opposed to only writing them to disk)?}
}
\value{
A named list containing
\itemize{
  \item{\code{models}  The PaleoCAR models, as computed by \code{\link{paleocar_models_batch}}.}
  \item{\code{recon}  The PaleoCAR reconstruction, as computed by \code{\link{predict_paleocar_models_batch}}.}
  \item{\code{errors}  The PaleoCAR reconstruction average LOOCV error, as computed by \code{\link{errors_paleocar_models_batch}}.}
  \item{\code{sizes}  The PaleoCAR model sizes, as computed by \code{\link{size_paleocar_models_batch}}.}
}
}
\description{
This is a wrapper function to the four primary methods in the PaleoCAR package, specifically for
processing predictands held in a RasterBrick. Additionally, it allows for the specification of
a \code{floor} and/or \code{ceiling} corrections for a reconstruction (e.g., of precipitation).
It returns a list containing the models, predictions, cross-validated errors, and model sizes.
}

