% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.paleoCAR.models.batch.R
\name{size.paleocar.models.batch}
\alias{size.paleocar.models.batch}
\title{Get a RasterBrick of PaleoCAR model sizes}
\usage{
size.paleocar.models.batch(models, prediction.years = NULL)
}
\arguments{
\item{models}{A PaleoCAR batch model, as returned from \link{paleoCAR.models.batch}.}

\item{prediction.years}{The set of years over which to generate model size rasters.}
}
\value{
A RasterBrick containing the model sizes per year.
}
\description{
This extracts the model size information 
from a set of PaleoCAR models and represents them as a RasterBrick with
layers for each reconstructed year.
}

