% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_paleocar_models_batch.R
\name{size_paleocar_models_batch}
\alias{size_paleocar_models_batch}
\title{Get a RasterBrick of PaleoCAR model sizes}
\usage{
size_paleocar_models_batch(models, prediction.years = NULL)
}
\arguments{
\item{models}{A PaleoCAR batch model, as returned from \link{paleocar_models_batch}.}

\item{prediction.years}{The set of years over which to generate model size rasters.}
}
\value{
A RasterBrick containing the model sizes per year.
}
\description{
This extracts the model size information 
from a set of PaleoCAR models and represents them as a RasterBrick with
layers for each reconstructed year.
}

