% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.mlm.R
\name{tidy_mlm}
\alias{tidy_mlm}
\title{Tidy function for mlm objects including confidence intervals}
\usage{
tidy_mlm(model, conf.int = FALSE, level = 0.95)
}
\arguments{
\item{model}{An object of class mlm.}

\item{conf.int}{Whether to calculate confidence intervals.}

\item{level}{Level for confidence interval}
}
\value{
A tibble of tidy model results.
}
\description{
Tidy function for mlm objects including confidence intervals
}
