% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_paleocar_models.R
\name{uncertainty_paleocar_models}
\alias{uncertainty_paleocar_models}
\title{Get reconstruction LOOCV uncertainty}
\usage{
uncertainty_paleocar_models(models, prediction.years = NULL, ...)
}
\arguments{
\item{models}{A PaleoCAR batch model, as returned from \link{paleocar_models}.}

\item{prediction.years}{The set of years over which to generate error estimates.}

\item{...}{Further arguments to be passed to other functions.}
}
\value{
The LOOCV statistics per year.
}
\description{
This extracts the leave-one-out cross-validation (LOOCV) uncertainty information 
from a set of PaleoCAR models and represents them as a RasterBrick with
layers for each reconstructed year. The LOOCV statistic is the sum of cross-validated 
squared errors for the model over the calibration period. This method returns the average uncertainty
by dividing the LOOCV statistic by the number of predictand observations, and then taking the square root.
Resulting uncertainty is delivered in the units of the predictand.
}

