% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPredlist.R
\name{getPredlist}
\alias{getPredlist}
\title{Get list of stepwise predictors from a reconstruction matrix}
\usage{
getPredlist(reconstruction.matrix)
}
\arguments{
\item{predictor.matrix}{A matrix of predictors for a reconstruction.
Columns are predictors, rows are years.}
}
\value{
A named list containing the names of available predictors
for the years beginning with the name of element "i" and
ending with element "i+1"-1.
}
\description{
This creates a list of available predictors along the years in
the reconstruction matrix.
}

