% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReconstructionMatrix.R
\name{getReconstructionMatrix}
\alias{getReconstructionMatrix}
\title{Get a matrix of predictors for a reconstruction}
\usage{
getReconstructionMatrix(chronologies, reconstruction.years = NULL,
  min.width = NULL)
}
\arguments{
\item{chronologies}{A matrix of predictors to be subset.
Columns are predictors, rows are years.}

\item{reconstruction.years}{A vector of years to be reconstructed.
If missing, will use all years in \code{chronologies}.}

\item{min.width}{integer, indicating the minimum number of tree-ring samples allowed for that year of a chronology to be valid.}
}
\value{
A matrix of predictors for a reconstruction.
}
\description{
This subsets a chronologies matrix to a set of reconstruction years.
}

