% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_paleocar_models.R
\name{predict_paleocar_models}
\alias{predict_paleocar_models}
\title{Get a RasterBrick reconstruction from a PaleoCAR batch model}
\usage{
predict_paleocar_models(models, prediction.years = NULL, ...)
}
\arguments{
\item{models}{A PaleoCAR batch model, as returned from \code{\link{paleocar_models}}.}

\item{prediction.years}{The set of years over which to generate reconstruction rasters. Optional.}

\item{...}{Further arguments to be passed to other functions.}
}
\value{
A tibble or RasterBrick containing the predictions for each year.
}
\description{
This generates a reconstruction RasterBrick from a PaleoCAR batch model.
}
