% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_paleocar_models.R
\name{uncertainty_paleocar_models}
\alias{uncertainty_paleocar_models}
\title{Get reconstruction standard errors of the models, through time}
\usage{
uncertainty_paleocar_models(models, prediction.years = NULL, ...)
}
\arguments{
\item{models}{A PaleoCAR batch model, as returned from \link{paleocar_models}.}

\item{prediction.years}{The set of years over which to generate error estimates.}

\item{...}{Further arguments to be passed to other functions.}
}
\value{
The standard error per year.
}
\description{
This extracts the standard errors 
from a set of PaleoCAR models and represents them as a RasterBrick with
layers for each reconstructed year.
}
